Moved FontStyle and FontColor to their own files
