+++ src/main/java/moe/maple/api/script/util/builder/StyleAndColorBuilder.java
@@ -149,58 +149,4 @@ public abstract class StyleAndColorBuilder<Builder extends StyleAndColorBuilder<
      * @return this
      */
     public abstract Builder get();
-
-    /**
-     * @author umbreon22
-     * String prefixes that adjust the style of client text.
-     */
-    public enum FontStyle {
-        NONE(""),
-        NORMAL("#n"),
-        BOLD("#e");
-        final String prefix;
-        FontStyle(String style) {
-            this.prefix = style;
-        }
-
-        /**
-         * Converts a Client Prefix into a Style
-         * @param p (Ex. #e)
-         * @return Color (Ex. would return BOLD)
-         */
-        public static FontStyle fromPrefix(String p) {
-            for(FontStyle style : values())
-                if(style.prefix.equals(p)) return style;
-            return FontStyle.NONE;
-        }
-    }
-
-    /**
-     * @author umbreon22
-     * String prefixes that adjust the color of client text.
-     */
-    public enum FontColor {
-        NONE(""),
-        BLACK("#k"),
-        BLUE("#b"),
-        RED("#r"),
-        PURPLE("#d"),
-        GREEN("#g");
-        final String prefix;
-        FontColor(String color) {
-            this.prefix = color;
-        }
-
-        /**
-         * Converts a Client Prefix into a Color
-         * @param p (Ex. #k)
-         * @return Color (Ex. would return BLACK)
-         */
-        public static FontColor fromPrefix(String p) {
-            for(FontColor color : values())
-                if(color.prefix.equals(p)) return color;
-            return FontColor.NONE;
-        }
-
-    }
 }
